[TurboQuant] Keep SoA routing and FP8 store semantics consistent

+++ vllm/v1/attention/ops/turboquant_soa/triton_turboquant_store.py
@@ -188,7 +188,12 @@ def _tq_fused_store_fp8(
     # ── FP8 KEY: cast to FP8 in-kernel and store at data_base ────────
     d_offs = tl.arange(0, BLOCK_D)
     d_mask = d_offs < D
-    k_vals = tl.load(Key_ptr + base + d_offs, mask=d_mask, other=0.0)
+    # Match the standard AoS store: promote before the FP8 conversion. This
+    # avoids backend-specific bf16/fp16 -> FP8 lowering differences, notably
+    # on ROCm targets.
+    k_vals = tl.load(Key_ptr + base + d_offs, mask=d_mask, other=0.0).to(
+        tl.float32
+    )
     k_fp8 = k_vals.to(tl.float8e4b15) if FP8_E4B15 else k_vals.to(tl.float8e4nv)
     k_bytes = k_fp8.to(tl.uint8, bitcast=True)
     tl.store(KV_cache_ptr + data_base + d_offs, k_bytes, mask=d_mask)
